Remove CompositeContentType
